// lib/alfredo/typed.d.ts
export interface AlfredMeta {
  name: string;
  version: string;
  uid: string;
  bundleId: string;
}

export interface AlfredEnv {
  version: string;
  theme: string;
  themeBackground: string;
  themeSelectionBackground: string;
  themeSubtext: number;
  data: string;
  cache: string;
  preferences: string;
  preferencesLocalHash: string;
}
// ...
export interface ScriptOutput {
  /**
   * Optional variables
   */
  variables?: {
    [key: string]: unknown;
  };
  /**
   * Items
   */
  items: ScriptItem[];
  /**
   * Rerun automatically after an interval
   *
   * Should be between ```0.1``` and ```5.0```
   */
  rerun?: number;
}